# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/sessions/{sessionId}/messages": {
      "get": {
        "operationId": "MessageController_getMessages",
        "parameters": [
          {
            "name": "sessionId",
            "required": true,
            "in": "path",
            "description": "Session ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "chatId",
            "required": false,
            "in": "query",
            "description": "Filter by chat ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "from",
            "required": false,
            "in": "query",
            "description": "Filter by sender. A phone also matches group messages via the author field and any lid that resolves to it.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Max messages to return (default 50)",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "offset",
            "required": false,
            "in": "query",
            "description": "Offset for pagination",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Message history"
          }
        },
        "summary": "Get message history for a session",
        "tags": [
          "messages"
        ]
      }
    },
    "/api/sessions/{sessionId}/messages/send-text": {
      "post": {
        "operationId": "MessageController_sendText",
        "parameters": [
          {
            "name": "sessionId",
            "required": true,
            "in": "path",
            "description": "Session ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendTextMessageDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Message sent",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Session not active or invalid request"
          },
          "404": {
            "description": "Session not found"
          }
        },
        "summary": "Send a text message",
        "tags": [
          "messages"
        ]
      }
    },
    "/api/sessions/{sessionId}/messages/send-template": {
      "post": {
        "operationId": "MessageController_sendTemplate",
        "parameters": [
          {
            "name": "sessionId",
            "required": true,
            "in": "path",
            "description": "Session ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendTemplateMessageDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Template rendered and sent",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Session not active or invalid request"
          },
          "404": {
            "description": "Session or template not found"
          }
        },
        "summary": "Render a stored text template and send it as a text message",
        "tags": [
          "messages"
        ]
      }
    },
    "/api/sessions/{sessionId}/messages/send-image": {
      "post": {
        "operationId": "MessageController_sendImage",
        "parameters": [
          {
            "name": "sessionId",
            "required": true,
            "in": "path",
            "description": "Session ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendMediaMessageDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Image sent",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Session not active or invalid request"
          }
        },
        "summary": "Send an image message",
        "tags": [
          "messages"
        ]
      }
    },
    "/api/sessions/{sessionId}/messages/send-video": {
      "post": {
        "operationId": "MessageController_sendVideo",
        "parameters": [
          {
            "name": "sessionId",
            "required": true,
            "in": "path",
            "description": "Session ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendMediaMessageDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Video sent",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Session not active or invalid request"
          }
        },
        "summary": "Send a video message",
        "tags": [
          "messages"
        ]
      }
    },
    "/api/sessions/{sessionId}/messages/send-audio": {
      "post": {
        "operationId": "MessageController_sendAudio",
        "parameters": [
          {
            "name": "sessionId",
            "required": true,
            "in": "path",
            "description": "Session ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendAudioMessageDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Audio sent",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Session not active or invalid request"
          }
        },
        "summary": "Send an audio/voice message",
        "tags": [
          "messages"
        ]
      }
    },
    "/api/sessions/{sessionId}/messages/send-document": {
      "post": {
        "operationId": "MessageController_sendDocument",
        "parameters": [
          {
            "name": "sessionId",
            "required": true,
            "in": "path",
            "description": "Session ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendMediaMessageDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Document sent",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Session not active or invalid request"
          }
        },
        "summary": "Send a document/file",
        "tags": [
          "messages"
        ]
      }
    },
    "/api/sessions/{sessionId}/messages/send-location": {
      "post": {
        "operationId": "MessageController_sendLocation",
        "parameters": [
          {
            "name": "sessionId",
            "required": true,
            "in": "path",
            "description": "Session ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendLocationDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Location sent",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponseDto"
                }
              }
            }
          }
        },
        "summary": "Send a location message",
        "tags": [
          "messages"
        ]
      }
    },
    "/api/sessions/{sessionId}/messages/send-contact": {
      "post": {
        "operationId": "MessageController_sendContact",
        "parameters": [
          {
            "name": "sessionId",
            "required": true,
            "in": "path",
            "description": "Session ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendContactDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Contact sent",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponseDto"
                }
              }
            }
          }
        },
        "summary": "Send a contact card message",
        "tags": [
          "messages"
        ]
      }
    },
    "/api/sessions/{sessionId}/messages/send-sticker": {
      "post": {
        "operationId": "MessageController_sendSticker",
        "parameters": [
          {
            "name": "sessionId",
            "required": true,
            "in": "path",
            "description": "Session ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendMediaMessageDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Sticker sent",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponseDto"
                }
              }
            }
          }
        },
        "summary": "Send a sticker message",
        "tags": [
          "messages"
        ]
      }
    },
    "/api/sessions/{sessionId}/messages/send-poll": {
      "post": {
        "operationId": "MessageController_sendPoll",
        "parameters": [
          {
            "name": "sessionId",
            "required": true,
            "in": "path",
            "description": "Session ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendPollDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Poll sent",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponseDto"
                }
              }
            }
          }
        },
        "summary": "Send a native WhatsApp poll",
        "tags": [
          "messages"
        ]
      }
    },
    "/api/sessions/{sessionId}/messages/reply": {
      "post": {
        "operationId": "MessageController_reply",
        "parameters": [
          {
            "name": "sessionId",
            "required": true,
            "in": "path",
            "description": "Session ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReplyMessageDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Reply sent",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponseDto"
                }
              }
            }
          }
        },
        "summary": "Reply to a message",
        "tags": [
          "messages"
        ]
      }
    },
    "/api/sessions/{sessionId}/messages/forward": {
      "post": {
        "operationId": "MessageController_forward",
        "parameters": [
          {
            "name": "sessionId",
            "required": true,
            "in": "path",
            "description": "Session ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ForwardMessageDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Message forwarded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponseDto"
                }
              }
            }
          }
        },
        "summary": "Forward a message to another chat",
        "tags": [
          "messages"
        ]
      }
    },
...
  },
  "info": {
    "title": "OpenWA API",
    "description": "Open Source WhatsApp API Gateway - Free, Self-Hosted HTTP API",
    "version": "0.13.0",
    "contact": {
      "name": "OpenWA",
      "url": "https://github.com/rmyndharis/OpenWA",
      "email": "yudhi@rmyndharis.com"
    }
  },
...
  "servers": [
    {
      "url": "http://{host}:{port}",
      "description": "OpenWA instance",
      "variables": {
        "host": {
          "default": "localhost"
        },
        "port": {
          "default": "2785",
          "description": "PORT env var"
        }
      }
    }
  ],
  "components": {
    "securitySchemes": {
      "X-API-Key": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      },
      "metrics-bearer": {
        "scheme": "bearer",
        "bearerFormat": "opaque",
        "type": "http",
        "description": "METRICS_TOKEN for the GET /api/metrics scrape endpoint"
      }
    },
    "schemas": {
...
      "SendTextMessageDto": {
        "type": "object",
        "properties": {
          "chatId": {
            "type": "string",
            "description": "WhatsApp chat ID (phone@c.us for individual, groupId@g.us for groups)",
            "example": "628123456789@c.us"
          },
          "text": {
            "type": "string",
            "description": "Text message content",
            "example": "Hello from OpenWA!",
            "maxLength": 4096
          },
          "mentions": {
            "description": "WIDs to @mention (e.g. [\"62811@c.us\"]). The text/caption must also contain the @<number> token.",
            "example": [
              "628123456789@c.us"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "chatId",
          "text"
        ]
      },
...
      "ReplyMessageDto": {
        "type": "object",
        "properties": {
          "chatId": {
            "type": "string"
          },
          "quotedMessageId": {
            "type": "string"
          },
          "text": {
            "type": "string",
            "maxLength": 4096
          }
        },
        "required": [
          "chatId",
          "quotedMessageId",
          "text"
        ]
      },
...
    }
  },
  "security": [
    {
      "X-API-Key": []
    }
  ]
}